requirements.txt
asgiref==3.7.2
django-jazzmin==2.6.0
pillow==10.2.0
sqlparse==0.4.4
